removing copied files
