--- stardis/opacities/__init__.py
@@ -1,1 +1,0 @@
-from stardis.opacities.base import *
