+++ stardis/radiation_field/opacities/opacities_solvers/voigt.py
@@ -1,9 +1,20 @@
 import numpy as np
 import numba
+from numba import cuda
+import cmath
+
+GPUs_available = cuda.is_available()
+
+if GPUs_available:
+    import cupy as cp
+
+SQRT_PI = np.sqrt(np.pi, dtype=float)
+SQRT_2 = np.sqrt(2, dtype=float)
+PI = float(np.pi)
 
 
-@numba.njit
-def faddeeva(z):
+@numba.njit()
+def _faddeeva(z):
     """
     The Faddeeva function. Code adapted from
     https://github.com/tiagopereira/Transparency.jl/blob/966fb46c21/src/voigt.jl#L13.
@@ -16,65 +27,90 @@ def faddeeva(z):
     -------
     w : complex
     """
-    s = abs(np.real(z)) + np.imag(z)
+    z = complex(z)
+    x = float(z.real)
+    y = float(z.imag)
+    t = y - 1j * x
+    s = abs(x) + y
+    w = complex(0.0)
+    u = t * t
 
-    if s > 15.0:
-        # Region I
-        w = 1j * 1 / np.sqrt(np.pi) * z / (z**2 - 0.5)
+    IN_REGION_I = s > 15.0
+    IN_REGION_II = (not IN_REGION_I) and (s > 5.5)
+    IN_REGION_III = (
+        (not IN_REGION_I) and (not IN_REGION_II) and (y >= 0.195 * abs(x) - 0.176)
+    )
+    IN_REGION_IV = (not IN_REGION_I) and (not IN_REGION_II) and (not IN_REGION_III)
 
-    elif s > 5.5:
-        # Region II
-        w = (
-            1j
-            * (z * (z**2 * 1 / np.sqrt(np.pi) - 1.4104739589))
-            / (0.75 + z**2 * (z**2 - 3.0))
+    # If in Region I
+    w = 1j * 1 / SQRT_PI * z / (z**2 - 0.5) if IN_REGION_I else w
+
+    # If in Region II
+    w = (
+        1j
+        * (z * (z**2 * 1 / SQRT_PI - 1.4104739589))
+        / (0.75 + z**2 * (z**2 - 3.0))
+        if IN_REGION_II
+        else w
+    )
+
+    # If in Region III
+    w = (
+        (16.4955 + t * (20.20933 + t * (11.96482 + t * (3.778987 + 0.5642236 * t))))
+        / (
+            16.4955
+            + t * (38.82363 + t * (39.27121 + t * (21.69274 + t * (6.699398 + t))))
         )
+        if IN_REGION_III
+        else w
+    )
 
-    else:
-        x = np.real(z)
-        y = np.imag(z)
-        t = y - 1j * x
-
-        if y >= 0.195 * abs(x) - 0.176:
-            # Region III
-            w = (
-                16.4955
-                + t * (20.20933 + t * (11.96482 + t * (3.778987 + 0.5642236 * t)))
-            ) / (
-                16.4955
-                + t * (38.82363 + t * (39.27121 + t * (21.69274 + t * (6.699398 + t))))
-            )
-
-        else:
-            # Region IV
-            u = t * t
-            numerator = t * (
-                36183.31
-                - u
-                * (
-                    3321.99
-                    - u
-                    * (
-                        1540.787
-                        - u * (219.031 - u * (35.7668 - u * (1.320522 - u * 0.56419)))
-                    )
-                )
-            )
-            denominantor = 32066.6 - u * (
-                24322.8
-                - u
-                * (
-                    9022.23
-                    - u * (2186.18 - u * (364.219 - u * (61.5704 - u * (1.84144 - u))))
-                )
-            )
-            w = np.exp(u) - numerator / denominantor
+    # If in Region IV
+    numerator = t * (
+        36183.31
+        - u
+        * (
+            3321.99
+            - u
+            * (1540.787 - u * (219.031 - u * (35.7668 - u * (1.320522 - u * 0.56419))))
+        )
+    )
+    denominator = 32066.6 - u * (
+        24322.8
+        - u
+        * (9022.23 - u * (2186.18 - u * (364.219 - u * (61.5704 - u * (1.84144 - u)))))
+    )
+    w = (cmath.exp(u) - numerator / denominator) if IN_REGION_IV else w
 
     return w
 
 
+@numba.vectorize(nopython=True)
+def faddeeva(z):
+    return _faddeeva(z)
+
+
+@cuda.jit
+def _faddeeva_cuda(res, z):
+    tid = cuda.grid(1)
+    size = len(res)
+
+    if tid < size:
+        res[tid] = _faddeeva(z[tid])
+
+
+def faddeeva_cuda(z, nthreads=256, ret_np_ndarray=False):
+    size = len(z)
+    nblocks = 1 + (size // nthreads)
+    z = cp.asarray(z, dtype=complex)
+    res = cp.empty_like(z)
+
+    _faddeeva_cuda[nblocks, nthreads](res, z)
+    return cp.asnumpy(res) if ret_np_ndarray else res
+
+
 @numba.njit
-def voigt_profile(delta_nu, doppler_width, gamma):
+def _voigt_profile(delta_nu, doppler_width, gamma):
     """
     Calculates the Voigt profile, the convolution of a Lorentz profile
     and a Gaussian profile.
@@ -94,6 +130,53 @@ def voigt_profile(delta_nu, doppler_width, gamma):
     phi : float
         Value of Voigt profile.
     """
-    z = (delta_nu + (gamma / (4 * np.pi)) * 1j) / doppler_width
-    phi = np.real(faddeeva(z)) / (np.sqrt(np.pi) * doppler_width)
+    delta_nu, doppler_width, gamma = float(delta_nu), float(doppler_width), float(gamma)
+
+    z = complex(delta_nu, gamma / (4 * PI)) / doppler_width
+    phi = _faddeeva(z).real / (SQRT_PI * doppler_width)
     return phi
+
+
+@numba.vectorize(nopython=True)
+def voigt_profile(delta_nu, doppler_width, gamma):
+    return _voigt_profile(delta_nu, doppler_width, gamma)
+
+
+@cuda.jit
+def _voigt_profile_cuda(res, delta_nu, doppler_width, gamma):
+    tid = cuda.grid(1)
+
+    size = len(res)
+
+    if tid < size:
+        res[tid] = _voigt_profile(delta_nu[tid], doppler_width[tid], gamma[tid])
+
+
+def voigt_profile_cuda(
+    delta_nu,
+    doppler_width,
+    gamma,
+    nthreads=256,
+    ret_np_ndarray=False,
+    dtype=float,
+):
+    arg_list = (
+        delta_nu,
+        doppler_width,
+        gamma,
+    )
+    shortest_arg_idx = np.argmin(map(len, arg_list))
+    size = len(arg_list[shortest_arg_idx])
+
+    nblocks = 1 + (size // nthreads)
+
+    arg_list = tuple(map(lambda v: cp.array(v, dtype=dtype), arg_list))
+
+    res = cp.empty_like(arg_list[shortest_arg_idx], dtype=dtype)
+
+    _voigt_profile_cuda[nblocks, nthreads](
+        res,
+        *arg_list,
+    )
+
+    return cp.asnumpy(res) if ret_np_ndarray else res
